pyproject.toml
[tool.poetry]
name = "build_sam_template"
version = "0.1.0"
description = ""
authors = ["Jeremy Rodrigues <jrodrigues@ripple.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
